FOSS-55 fixed login issue
The issue was caused by a reload by Shopware after the user got redirected from `/admin?state=...` to `/admin?state=...#/login`. In some cases in the time between the new route was set by VueJS and the page got reloaded, the token request was already sent to Shopware but the response was not processed by the VueJS application, as the page reload interrupts this. However, part of the token retrieval is that the used state gets invalidated to prevent reuse. After the page reloaded, the state was no longer valid causing the login process to terminate.

This commit also relates to: #26, #28, #31